requirements.txt
# 1. Frameworks Principales
streamlit
llama-index
llama-index-core

# 2. Conectores de Vector Stores (Base de Datos)
pinecone
llama-index-vector-stores-pinecone

# 3. Modelos de Embeddings Gratuitos (Hugging Face)
sentence-transformers
llama-index-embeddings-huggingface

# 4. Conectores de LLM Gratuitos (Para la Generación de Respuestas)
llama-index-llms-huggingface

# 5. Utilidades y Procesamiento de Documentos
python-dotenv
pypdf